+++ lwk_bindings/src/electrum_client.rs
@@ -3,7 +3,9 @@ use std::time::Duration;
 
 use lwk_wollet::clients::blocking::BlockchainBackend;
 
-use crate::{BlockHeader, LwkError, TokenProvider, Transaction, Txid, Update, Wollet};
+use crate::{
+    BlockHeader, LwkError, TokenProvider, Transaction, Txid, Update, Wollet, WolletDescriptor,
+};
 
 /// A client to issue TCP requests to an electrum server.
 #[derive(uniffi::Object, Debug)]
@@ -152,6 +154,15 @@ impl ElectrumClient {
     pub fn tip(&self) -> Result<Arc<BlockHeader>, LwkError> {
         let tip = self.inner.lock()?.tip()?;
         Ok(Arc::new(tip.into()))
+    }
+
+    /// Whether the descriptor has any tx using the first `gap_limit` addresses (default 20)
+    pub fn has_txs(
+        &self,
+        descriptor: &WolletDescriptor,
+        gap_limit: Option<u32>,
+    ) -> Result<bool, LwkError> {
+        Ok(self.inner.lock()?.has_txs(descriptor.as_ref(), gap_limit)?)
     }
 }